Scrolling to bottom if the messages are overflowing
Recovering the feature from the example, but with a difference. In the example the scrollTo is implemented using `useEffect`, here we are using `refCallbacks`. 

refCallbacks are always, mount, rerenders and unmount.

The benefit against useEffect, we don't need two hooks and the behavior is collocated in the component that affects, instead of a random useEffect.